build: adapt /src layout
From: https://packaging.python.org/en/latest/discussions/src-layout-vs-flat-layout/

The src layout involves an additional step in the development workflow of a
project (typically, an editable installation is used for development and a
regular installation is used for testing).

(...)

The src layout helps enforce that an editable installation is only able to
import files that were meant to be importable.

(...)

This is especially relevant when the editable installation is implemented
using a path configuration file that adds the directory to the import path.